requirements.txt
certifi==2021.10.8
chardet==3.0.4
click==8.0.4
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.1.0
numpy==1.22.3
pandas==1.4.1
Pillow==9.0.1
python-dateutil==2.8.2
pytube==12.0.0
pytz==2021.3
rfc3986==1.5.0
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SpeechRecognition==3.8.1
threadpoolctl==3.1.0
tqdm==4.63.0
